tangle notebook rendering common.
